take a big leap after some life network testing
Highlights:
 outgoing compression stream & UDP
 tuned a lot of logging down.
 better and fixed udp reassambly
 removal of libxml, dependencies stink!
 lot of crash/corruption/whatever fixes

This warants a new version: 0.0.00.11
"And there was great rejoice!"

since we are now doing more cpy/move within recv buffers, and other stuff,
new in our list of brocken lib functions:
added:
 lib/memmove.c
 lib/generic/memmove.c
	add a memmove to our toys, we can forward most of the time to our
	 memcpy, which we know will copy forward
 lib/memcpy.c
 lib/x86/memcpy.c
 lib/x86/memcpy_tmpl.c
	add an special x86 memcpy to be not compiler dependent
 lib/guid.c
 lib/guid.h
	foo to generate valid guids, maybe move some stuff here
 lib/entities.c
	add a function to unencode html entities, now that libxml is gone

Other advances, small, big, or fixes for brocken stuff:
 G2Acceptor.c
	repair given status
	handshake compression with ultra peers, in & out, now we have that
	make sure hubs which where hubs already at handshake time were propaly recorded
 G2ConHelper.c
	make sure not to overwrite poll interrests in passed struct
	make sure to only epoll_ctl if connection is not oneshot
 G2ConRegistry.h
 G2ConRegistry.c
	add iterator for all connections
	do more cleanup at shutdown
 G2Connection.h
 G2Connection.c
	add more header fields
	give internal alloc pool a kill switch
	remove the fixed z_stream struct from connection to save those bytes
	-> bringing g2_connection_t under 512 byte
	reorg cleanup and freeing, make sure qht gets put
	add z timeout
	reorg remote_ip and listen_ip
 G2GUIDCache.h
	depend on new guid stuff
 G2Handler.c
	add outgoing compression
	reomve dead variable
	change for z_stream separation
	add comment for LZO
 G2KHL.c
	add important TODO, target for next version
 G2MainServer.c
	remove libxml init
	add dump signal
	add more vars and config foo, typos
	periodically send a HAW
 G2MainServer.h
	remove dead stuff
	add more config fields
	revamp local time system by introducing a master time
 G2Packet.h
	give packet sorting some real life numbers
	add more packet types
 G2Packet.c
	likewise, logg the farther packet now we have that
	SNAs can contain a port...
	add /KHL/NH
	fix a missed out_free
	add a function to generate & send a HAW, use out new guid generate stuff
	comment more, rename some variables, fix some missed ->is_bigendian
	give packet allocator a kill switch
	 -> it is brocken ATM, some list coruption after some time
	add a packet statistic dump, for debug only, not SMP save
	forward more packets
 G2PacketTyperGenerator.c
	now we have life sort values, they can get big, make the acc larger
 G2PacketSerializer.c
	give the extrator a switch to force extraction if data is not already
	 completly in the buffer
	smelled like some state corruption
 G2QHT.c
	throw out lib xml parsing
	fix some search_add_* to correctly initialize the string
	fix access to uninitialized qh-tables
	use spedial bitfield_and in the right cases
	remove the compressed size test, it fired, and fired, and...
 G2QueryKey.c
	kick init & rekeying of guid generation
 G2UDP.c
	make reassambly cache much larger
	remove part count restriction
	use a list stolen from packet space instead of an array
	copy buffers together
	reuse the buffer space for packet data if it fits
	fix a truckload of memory leaks
	implement outgoing compression
	give udp dumping a kill switch
	prevent sending to reserved IPs or when no dst port is given

 builtin_defaults.h
	adjust defaults
	make norm_buffer smaller to take header into account, so they fit
	 in exactly 4k
 gup.h
 gup.c
	remove locking foo, try to lock in kernel
	add a z_flush timeout function to retrigger write attempt
 timeout.c
	maintain the master clock
	reuse last gettime value, less syscalls

 lib/backtrace.c
	also handle SIGABRT, to not miss a backtrace in that case
 lib/bitfield_rle.c
	fix real thinko, access over array
	add function to directly and while uncompressing (which is
	easy, our compression format saves the places which need
	an and to do something special...)
 lib/combo_addr.h
	fix wrong net match (ups...)
	add test for reserved/forbidden IPs
 lib/config_parser.h
 lib/config_parser.c
	add function to assign a string value
 lib/hzp.c
	fix nasty bug, codepath was never executed before,
	 crashed on first use...
 lib/recv_buff.c
	add kill switch for local allocator
	make sure buffers handed out are proper initialized
 lib/hthash.h
	add jhash for multiplication challanged archs, prototype
 lib/rbtree.c
	add annotation
 lib/tchar.c
	fix a nasty crash on utf-8 conversion, 3-star programming
	 gone wrong...
 lib/vsnprintf.c
	print guids more guid like
 lib/generic/memcpy.c
	make sure our memcpy always copies forward
 lib/my_epoll_devpoll.c
 lib/my_epoll_eport.c
 lib/my_epoll_kqueue.c
 lib/my_epoll_poll.c
	now that gup does not lock, make sure locking is in place
 lib/sparc/memcpy.c
	use more lib foo
 lib/x86/memand_tmpl.c
 lib/x86/memneg_tmpl.c
 lib/x86/memxorcpy.c
 lib/x86/memxorcpy_tmpl.c
	remove test
 lib/x86/x86.h
	add another load case
 lib/x86/my_bitops.c
	typo
 lib/x86/tstrchrnul.c
	fix asm
 lib/my_bitops.h
 lib/module.make
	knot in new toys
 Makefile
 README
	remove libxml

+++ G2PacketTyperGenerator.c
@@ -2,19 +2,19 @@
  * G2PacketTyperGenerator.c
  * Automaic generator for the G2-Packet typer tables
  *
- * Copyright (c) 2008 Jan Seiffert
+ * Copyright (c) 2008-2010 Jan Seiffert
  *
  * This file is part of g2cd.
  *
  * g2cd is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License version
  * 2 as published by the Free Software Foundation.
- * 
+ *
  * g2cd is distributed in the hope that it will be useful, but
  * WITHOUT ANY WARRANTY; without even the implied warranty of
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
  * GNU General Public License for more details.
- * 
+ *
  * You should have received a copy of the GNU General Public
  * License along with g2cd; if not, write to the Free Software
  * Foundation, Inc., 59 Temple Place, Suite 330, Boston,
@@ -35,7 +35,7 @@
 struct p_names
 {
 	const char c[8];
-	const int weight;
+	const long long weight;
 };
 
 #define ENUM_CMD(x, y) { .c = str_it(x), .weight = (y) }
@@ -52,7 +52,7 @@ struct tree_hl
 	char c;
 	bool last;
 	enum g2_ptype type;
-	int weight;
+	long long weight;
 	int line;
 	size_t num_child;
 	size_t size_child;
@@ -152,7 +152,7 @@ int main(int argc, char *argv[])
 	return EXIT_SUCCESS;
 }
 
-static void make_tree_hl_name_r(const char *name, struct tree_hl *t, int weight, enum g2_ptype type)
+static void make_tree_hl_name_r(const char *name, struct tree_hl *t, long long weight, enum g2_ptype type)
 {
 	struct tree_hl *x;
 	int i;
@@ -195,7 +195,7 @@ recurse:
 	}
 }
 
-static void make_tree_hl_name(const char *name, int weight, enum g2_ptype type)
+static void make_tree_hl_name(const char *name, long long weight, enum g2_ptype type)
 {
 	struct tree_hl *t;
 	unsigned c = name[0];
@@ -270,7 +270,7 @@ static void clean_tree_hl(void)
 
 static void print_tree_hl_r(struct tree_hl *t, int level)
 {
-	printf("%c %i%s\t", t->c ? t->c : '0', t->weight, t->last ? "!" : "");
+	printf("%c %lli%s\t", t->c ? t->c : '0', t->weight, t->last ? "!" : "");
 
 	if(!t->last)
 	{
